WignerSeitz

Documentation for WignerSeitz.

WignerSeitz.reduce_basis!

WignerSeitz.reduce_basis! — Method

reduce_basis!(M::AbstractMatrix{<:Real})

Reduces the basis vectors associated with a lattice to a form where all lattice vectors have minimum length and all angles between lattice vectors are at least 90 degrees. This is the form of the "superbase" described in Conway and Sloane's 1992 paper.

Essentially, this can be thought of as a "QRU" decomposition, where Q is a point isometry (rotation/reflection/etc.), R is an upper triangular matrix (this essentially reorients the basis vectors of the space), and U is a unimodular matrix (all lattices whose basis vectors are transformed by a unimodular matrix are identical), which permutes the columns of R, In this case, we minimize the dot product ofeach column of R with itself by right multiplication with a suitable unimodular matrix.