WignerSeitz

Documentation for WignerSeitz.

WignerSeitz.HermiteReducedLattice

WignerSeitz.LatticeReduction

WignerSeitz.gram

WignerSeitz.hermite_reduce

WignerSeitz.hermite_reduce!

WignerSeitz.isobtuse

WignerSeitz.projection

WignerSeitz.projection

WignerSeitz.selling

WignerSeitz.superbase

WignerSeitz.superbase_vector

WignerSeitz.HermiteReducedLattice — Type

HermiteReducedLattice{T,M<:AbstractMatrix{T}} <: LatticeReduction{T,M}

Stores a Hermite reduced basis B and the associated unimodular factor U such that for an input matrix M, L = U*M. This is an exactly reduced lattice.

WignerSeitz.LatticeReduction — Type

LatticeReduction{T,M<:AbstractMatrix{T}} <: LinearAlgebra.Factorization{T}

Supertype for the results of lattice reductions: algorithms which convert an m×n matrix (which represents an n-dimensional lattice in an m-dimensional space) to a sufficiently short basis.

WignerSeitz.gram — Method

WignerSeitz.gram(M::AbstractMatrix) -> typeof(M)

Returns the Gram matrix of M.

WignerSeitz.hermite_reduce! — Method

hermite_reduce!(M::AbstractMatrix) -> HermiteReducedLattice{eltype{M},M}

Performs an exact lattice reduction (in other words, without approximations to perform the reduction in polynomial time) of a lattice whose basis vectors are the columns of M. This lattice reduction occurs in-place, and the result is returned as a HermiteReducedLattice.

For a lattice reduction which creates a copy of the original matrix, or lattice reduction of an SMatrix, see hermite_reduction.

WignerSeitz.hermite_reduce — Method

hermite_reduce(M::AbstractMatrix) -> HermiteReducedLattice{eltype{M},M}

Performs an exact lattice reduction (in other words, without approximations to perform the reduction in polynomial time) of a lattice whose basis vectors are the columns of M. This creates a copy of the input matrix and modifies that copy in-place.

WignerSeitz.isobtuse — Method

WignerSeitz.isobtuse(M::AbstractMatrix) -> Bool

Returns true if a lattice basis is obtuse, meaning that all of the Selling parameters of the lattice basis are positive.

WignerSeitz.projection — Method

WignerSeitz.projection(M::AbstractMatrix, a, b) -> Number

Calculates the projection of M[:,a] onto M[:,b].

WignerSeitz.projection — Method

WignerSeitz.projection(u::AbstractVector, v::AbstractVector) -> Number

Calculates the projection of u onto v: dot(u,v) / dot(v,v).

WignerSeitz.selling — Method

WignerSeitz.selling(M::AbstractMatrix) -> typeof(M)

Returns the Selling matrix of M, which is equal to the negative Gram matrix, but includes an extra column for the superbase vector.

WignerSeitz.superbase — Method

WignerSeitz.superbase(M::AbstractMatrix) -> AbstractMatrix

Constructs the superbase of M, which is equal to M with an additional column containing the negative sum of all basis vector (columns) combined: -sum(M[:,n] for n in axes(M,2)).

WignerSeitz.superbase_vector — Method

WignerSeitz.superbase_vector(M::AbstractMatrix) -> AbstractVector

Returns the superbase vector of basis M, the negative sum of all basis vector (columns) combined: -sum(M[:,n] for n in axes(M,2)).